Update message/translation libraries to API 8
These haven't been tested yet, and there are likely some issues arising from the
translation. There may also be a better approach using TranslatableComponent,
which would allow it to be sent directly without the helper functions to get the
source's locale.

Furthermore, this doesn't change the behavior of these classes from FlashLibs.
It is likely the argument format will change from that later on (currently is
${style?@key}, likely going to just ${key} and keeping the style out of the
argument - convenient, but messes with readability and the implementation is
harder because there's no provided style parameter).

Finally, these classes will eventually be a part of FlashLibs for API 8, but for
now it's easier to work with them under this responsitory and then link back to
the commits when it's time to move them over to preseve the history by some
degree.