Refactor project to target Sponge API-7
This had two main components. First, the Gradle workflow needed to be
changed to earlier versions to ensure the mcmod.info file was properly
generated. Second, the implementation itself had to be manually updated
for API-7, which was generally pretty straightforward. The only notable
difference is that since MC 1.12.2 still has item data, the ITEM_TYPE
serializer was updated to support a `/<data>` suffix after the id.

Additionally, since FlashLibs already targets API-7 the corresponding
services added for API-8 can be removed in favor of the library itself.