Add message/broadcast for crates/rewards
The configuration behavior is that if either the crate/reward has a
message defined (priority to the reward) and neither of the messages are
empty, it is sent to the player. This allows crates to define a default
message that can be either overridden or silenced by the reward, while
rewards can also have messages defined silenced by a given crate.