Add basic effects (firework, potion, sound)
Effects are attached to the crate and can be triggered by a variety of
actions (idle not yet implemented). Effects are either player-based
(potion) or location-based (firework/sound), which incorporate a target
and offset into the reference value. Unlike other effects/components,
fireworks have undefined values selected randomly for variety.